# Flexbox: fix performance of nodes with very large numbers of children