# Objective

This is huge performance win on our 100k node "wide" benchmark, bringing the numbers roughly in line with our "deep" benchmarks for the same node count. Somewhat embarrassingly this one line was dominating performance for very wide trees:

```rust
let order = tree.children(node).position(|n| n == item.node).unwrap() as u32;
```

## Benchmark Results

| Benchmark          | Node Count | Depth | Yoga ([ba27f9d]) | Taffy 0.3     | Taffy (main)  | Taffy (this PR) |
| ---                | ---        | ---   | ---              | ---           | ---           | ---             |
| yoga 'huge nested' | 1,000      | 3     | 408.72 µs        | 329.04 µs     | 177.16 µs     | 295.64 µs       |
| yoga 'huge nested' | 10,000     | 4     | 3.9276 ms        | 4.3486 ms     | 2.8241 ms     | 2.7671 ms       |
| yoga 'huge nested' | 100,000    | 5     | 46.782 ms        | 38.559 ms     | 29.523 ms     | 24.912 ms       |
| big trees (wide)   | 1,000      | 1     | 741.25 µs        | 505.99 µs     | 464.12 µs     | 486.94 µs       |
| big trees (wide)   | 10,000     | 1     | **7.1613 ms**    | **8.3395 ms** | **6.5570 ms** | **5.1362 ms**   |
| big trees (wide)   | 100,000    | 1     | **132.09 ms**    | **247.42 ms** | **196.79 ms** | **64.096 ms**   |
| big trees (deep)   | 4,000      | 12    | 2.2308 ms        | 1.7400 ms     | 1.7071 ms     | 1.8021 ms       |
| big trees (deep)   | 10,000     | 14    | **5.8912 ms**    | **4.4445 ms** | **4.3659 ms** | **4.4437 ms**   |
| big trees (deep)   | 100,000    | 17    | **76.437 ms**    | **63.778 ms** | **60.390 ms** | **62.042 ms**   |
| super deep         | 1,000      | 1,000 | 561.34 µs        | 472.85 µs     | 412.16 µs     | 415.73 µs       |


## Context

Discovered while trying to optimise `display: contents` which makes calling `tree.children(node)` more expensive.